EfSectionOption is set only for specific sections
Sections where the ef_section_options.section is enforced are:
:DesignDetail, :ArmDetail, :BaselineCharacter, :OutcomeDetail

This is because we do not support by_section mass import yet.
AdverseEvents is the exception. AdverseEvents are always by_arm
and the importer expects an arm and arm description column. The
problem is that the ef_section_options.section by_arm value was
set to 0 even during AdverseEvents imort. This fixes this.