Subquestions can now be saved for existing questions.
The logic in DesignDetailDataPoint model was to check if
the subquestion value exists in the subquestion hash. This
is faulty since multiple questions could potentially have
the same subquestion value entered.

Also in the case of nested subquestion hashes the logic
always returned true in some cases and so no new sub-
question values could be added or changed once the
question was saved.

The current logic forgoes any matching and simply replaces
the subquestion for the matching datapoint entries.