Fixes a bug where all three buttons submit.
In diagnostic test result table regardless of which button was pressed,
the form was submitted. This only happens in IE 11 and is caused because
all three buttons are defined as submit buttons, even though only 1 is
meant to submit and the other two have on click event handlers via
javascript.